+++ internal/table/keep_in_cache.go
@@ -2,7 +2,6 @@ package table
 
 import "github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Table"
 
-func keepInCache(req *Ydb_Table.ExecuteDataQueryRequest) bool {
-	p := req.QueryCachePolicy
+func keepInCache(p *Ydb_Table.QueryCachePolicy) bool {
 	return p != nil && p.KeepInCache
 }