+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Supported internal allocator in `{session,statement}.Execute` for decrease memory usage
+
 ## v3.40.1
 * Added constructor of `options.TimeToLiveSettings` and fluent modifiers